+++ changelog.md
@@ -10,6 +10,7 @@ vX.X.X
 - Moved interface files from "interfaces/" and into the package/folder that they have most relevance with, closes #121.
 - Added total TX/RX byte counts to the bottom status bar, closes #123.
 - Added maximum flash rate functionality to the LED indicators, closes #118.
+- Added ability for logging engine to "swallow" ANSI escape codes, closes #125.
 
 v0.6.2
 ------